新的开始

+++ anchongnet/app/Http/Controllers/admin/goodController.php
@@ -148,6 +148,7 @@ class goodController extends Controller
             [
                 'cat_id'=>$catid,
                 'goods_id'=>$request->name,
+                //商品规格（属性)
                 'goods_name'=>trim($spetag),
                 'model' => $request->model,
                 'market_price'=>$request->marketprice,
@@ -208,8 +209,6 @@ class goodController extends Controller
        foreach($arr_key as $k) {
            DB::insert("insert into anchong_goods_suggestion (`str`) values ('$k') on duplicate key update snums=snums+1");
        }
-
-       /**清除关键字缓存操作*/
 
         /**
          * 向仓库表中插入
@@ -373,6 +372,10 @@ class goodController extends Controller
             $data = DB::table('anchong_goods_specifications')->where('gid',$aid)->get();
             if (!$data || Gate::denies('shopres',$data)) {
                 return null;
+            }
+            //促销中不可删除
+            if ($data[0]->promotion_price > 0 ) {
+                return '促销中的商品不可删除';
             }
             DB::beginTransaction();
             $res['spe'] = DB::table('anchong_goods_specifications')->where('gid',$aid)->delete();
